+++ flutter_front/lib/common/styles/sizes.dart
@@ -1,28 +1,29 @@
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+
 const double kLayoutMaxSize = 1440;
 const double kLayoutMarginSize = 36.0;
 const double kLayoutGutterSize = 12.0;
 
-const double kTextTitleSize = 30.0;
-const double kTextLargeSize = 28.0;
-const double kTextMiddleSize = 21.0;
-const double kTextSmallSize = 18.0;
-const double kTextMiniSize = 14.0;
+double kTextTitleSize = 30.0.sp;
+double kTextLargeSize = 28.0.sp;
+double kTextMiddleSize = 21.0.sp;
+double kTextSmallSize = 18.0.sp;
+double kTextMiniSize = 14.0.sp;
+// const double kTextTitleSize = 25.0;
+// const double kTextLargeSize = 21.0;
+// const double kTextMiddleSize = 18.0;
+// const double kTextSmallSize = 14.0;
+// const double kTextMiniSize = 10.0;
 
-const double kPaddingXLargeSize = 32.0;
-const double kPaddingLargeSize = 16.0;
-const double kPaddingMiddleSize = 12.0;
-const double kPaddingSmallSize = 8.0;
-const double kPaddingMiniSize = 4.0;
+double kPaddingXLargeSize = 32.0.w;
+double kPaddingLargeSize = 16.0.w;
+double kPaddingMiddleSize = 12.0.w;
+double kPaddingSmallSize = 8.0.w;
+double kPaddingMiniSize = 4.0.w;
 
-const double kIconLargeSize = 41.0;
-const double kIconMiddleSize = 28.0;
-const double kIconSmallSize = 20.0;
-const double kIconMiniSize = 16.0;
+double kIconLargeSize = 41.0.w;
+double kIconMiddleSize = 28.0.w;
+double kIconSmallSize = 20.0.w;
+double kIconMiniSize = 16.0.w;
 
 const double kBorderRadiusSize = 10.0;
-
-const double kListViewTileHeightSize = 100.0;
-
-const double kMainPageCalendarHeightSize = 833;
-
-const double kBorderSideWidth = 0.5;
