+++ flutter_front/lib/common/view/construction_screen.dart
@@ -18,12 +18,12 @@ class ConstructionScreen extends StatelessWidget {
             width: 300,
             height: 300,
           ),
-          const Text(
+          Text(
             "해당 기능은 공사중 입니다.\n빠른시일 내에 오픈하겠습니다.",
             style: kTextMainStyleLarge,
             textAlign: TextAlign.center,
           ),
-          const SizedBox(height: kPaddingXLargeSize),
+          SizedBox(height: kPaddingXLargeSize),
         ],
       ),
     );
