+++ flutter_front/lib/reservation_status/viewmodel/pre_reservation_viewmodel/progress_reservation_viewmodel.dart
@@ -47,7 +47,7 @@ class ProgressReservationViewModel extends ChangeNotifier {
 
     CustomDialogUtil.showCustomDialog(
         dialog: CustomDialog(
-          title: const Text(
+          title: Text(
             "우선예약 중단 확인",
             style: kTextMainStyleMiddle,
           ),
@@ -75,7 +75,7 @@ class ProgressReservationViewModel extends ChangeNotifier {
 
     CustomDialogUtil.showCustomDialog(
         dialog: CustomDialog(
-          title: const Text(
+          title: Text(
             "우선예약 재개 확인",
             style: kTextMainStyleMiddle,
           ),
