+++ flutter_front/lib/slack_message/view/message_view.dart
@@ -26,7 +26,7 @@ class MessageView extends ConsumerWidget {
           color: kMainColor,
           borderRadius: BorderRadius.circular(kBorderRadiusSize),
         ),
-        child: const Column(
+        child: Column(
           crossAxisAlignment: CrossAxisAlignment.stretch,
           children: [
             Text(
